5. Метод по адресу /calculator/divide?num1=5&num2=5 должен разделить num1 на num2 и вернуть результат в формате “5 / 5 = 1”.

+++ src/main/java/weblib/hw2_4/CalcService.java
@@ -7,6 +7,6 @@ public interface CalcService {
 
     int multiply(int num1, int num2);
 
-    int divide(int num1, int num2);
+    double divide(int num1, int num2);
 
 }
